Resolve AL2 pip package path issues (#282)
* Resolve AL2 pip package path issues


* Add small tweaks


---------

+++ packer/scripts/al2/al2-agent-setups.sh
@@ -8,7 +8,11 @@
 
 set -ex
 
-whoami
+CURR_USER="ec2-user"
+if [ `whoami` != "$CURR_USER" ]; then
+    echo "You must run this script on AL2 AMI with 'ec2-user' as user, exit 1"
+    exit 1
+fi
 
 sudo yum clean all
 sudo rm -rf /var/cache/yum /var/lib/yum/history
@@ -19,18 +23,21 @@ sudo amazon-linux-extras install java-openjdk11 -y
 sudo yum install -y which curl git gnupg2 tar net-tools procps-ng python3 python3-devel python3-pip zip unzip jq
 sudo yum install -y docker ntp
 sudo yum groupinstall -y "Development Tools"
-sudo ln -sfn `which pip3` /usr/bin/pip && sudo pip3 install pipenv awscli docker-compose && sudo ln -sfn ~/.local/bin/pipenv /usr/local/bin
+
+curl -o- https://bootstrap.pypa.io/get-pip.py | python3
+pip install pipenv awscli docker-compose && echo "export PATH=$PATH:$HOME/.local/bin" >> $HOME/.bashrc
+pipenv --version && aws --version && docker-compose --version
 
 sudo sed -i 's/OPTIONS/# OPTIONS/g' /etc/sysconfig/docker
 cat /etc/sysconfig/docker
 sudo systemctl restart docker && sudo systemctl enable docker && sudo systemctl status docker
 sudo systemctl restart ntpd && sudo systemctl enable ntpd && sudo systemctl status ntpd
-sudo usermod -a -G docker `whoami`
+sudo usermod -a -G docker $CURR_USER
 
 sudo yum-config-manager --add-repo https://cli.github.com/packages/rpm/gh-cli.repo
 sudo yum install -y gh
 
 sudo yum clean all
 
-sudo mkdir -p /var/jenkins && sudo chown -R ec2-user:ec2-user /var/jenkins
+sudo mkdir -p /var/jenkins && sudo chown -R $CURR_USER:$CURR_USER /var/jenkins
